moodboard-project/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
rq==1.15.1
python-dotenv==1.0.0
cryptography==41.0.7
pyjwt==2.8.0
pillow==10.1.0
aiofiles==23.2.1
structlog==23.2.0
gunicorn==21.2.0
bandit==1.7.5
safety==2.3.5